64-bit DeterministicRandom() seed and getRandomSeed()
DeterministicRandom() can use a boost::random::mt19937_64
to take a 64-bit seed and return 64-bit random values,
and this simplifies things because fdb always wants 64 bits.
Performance should be slightly better because the 32-bit
mt199937 was always called twice to get 64 bits.

Also fixes apple/foundationdb#11446 "idempotency IDs have insufficient entropy"

+++ fdbserver/SimulatedCluster.cpp
@@ -2658,7 +2658,7 @@ void setupSimulatedSystem(std::vector<Future<Void>>* systemActors,
 using namespace std::literals;
 
 // Actor that waits for a specified simulation time and then resets the random seed
-Future<Void> reseedRandomAtTime(double waitTime, uint32_t newSeed) {
+Future<Void> reseedRandomAtTime(double waitTime, uint64_t newSeed) {
 	co_await delay(waitTime);
 	TraceEvent("ResettingRandomSeed").detail("WaitTime", waitTime).detail("NewSeed", newSeed);
 	deterministicRandom()->resetSeed(newSeed);
@@ -2813,7 +2813,7 @@ static Future<Void> simulationSetupAndRunImpl(std::string dataFolder,
 		// If reseedTime is set, schedule a random seed reset at a random time between [0, reseedTime]
 		if (reseedTime >= 0.0) {
 			double actualReseedTime = nondeterministicRandom()->random01() * reseedTime;
-			uint32_t newSeed = platform::getRandomSeed();
+			uint64_t newSeed = platform::getRandomSeed();
 			TraceEvent("SchedulingRandomSeedReset")
 			    .detail("ReseedTimeMax", reseedTime)
 			    .detail("ActualReseedTime", actualReseedTime)
